+++ services/user-service/src/UserService.Api/Program.cs
@@ -62,9 +62,12 @@ using (var scope = app.Services.CreateScope())
 
 app.UseMiddleware<CorrelationIdMiddleware>();
 app.UseMiddleware<GlobalExceptionMiddleware>();
+app.UseRouting();
+app.UseHttpMetrics();
 app.UseAuthentication();
 app.UseAuthorization();
 app.MapControllers();
+app.MapMetrics();
 app.MapGrpcService<UserGrpcService>();
 app.MapHealthChecks("/health", new HealthCheckOptions
 {